+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on Keep a Changelog and this project adheres to Semantic Ver
 ## [1.5.2] - 2026-03-30
 ### Added
 - Add sorting options for the Network Health Scanner display lists, allowing sorting by distance or name. Chokepoints still sort by excess channels by default, sorting by distance/name as a tiebreaker. Sorting preference is saved per-tab.
+- Right-clicking the AutoCrafter with a crafting pattern in hand will insert it into the first available slot. Opens the GUI normally if the pattern is invalid (processing) or the crafter is full.
+
+### Fixed
+- Fix Network Health Scanner not resizing GUI when opened just after a scan (or as the scan is going).
 
 
 ## [1.5.1] - 2026-03-27